flatten proto components
