+++ packages/insomnia/src/ui/components/proto-file/proto-file-list.tsx
@@ -1,16 +1,25 @@
 import React, { FunctionComponent } from 'react';
+import styled from 'styled-components';
 
 import { ProtoDirectory } from '../../../models/proto-directory';
 import type { ProtoFile } from '../../../models/proto-file';
 import { ListGroup, ListGroupItem } from '../list-group';
-import { ProtoDirectoryListItem } from './proto-directory-list-item';
-import { ProtoFileListItem } from './proto-file-list-item';
+import { Button } from '../themed-button';
 
 export type SelectProtoFileHandler = (id: string) => void;
 export type DeleteProtoFileHandler = (protofile: ProtoFile) => void;
 export type DeleteProtoDirectoryHandler = (protoDirectory: ProtoDirectory) => void;
 export type UpdateProtoFileHandler = (protofile: ProtoFile) => Promise<void>;
 export type RenameProtoFileHandler = (protoFile: ProtoFile, name?: string) => Promise<void>;
+export const ProtoListItem = styled(ListGroupItem).attrs(() => ({
+  className: 'row-spaced',
+}))`
+  button i.fa {
+    font-size: var(--font-size-lg);
+  }
+
+  height: var(--line-height-sm);
+`;
 
 export interface ExpandedProtoDirectory {
   files: ProtoFile[];
@@ -35,25 +44,70 @@ const recursiveRender = (
   handleDeleteDirectory: DeleteProtoDirectoryHandler,
   selectedId?: string,
 ): React.ReactNode => ([
-  dir && (<ProtoDirectoryListItem
-    key={dir.name}
-    dir={dir}
-    indentLevel={indent++}
-    handleDeleteDirectory={handleDeleteDirectory}
-  />),
+  dir && (
+    <ProtoListItem indentLevel={indent}>
+      <span className="wide">
+        <i className="fa fa-folder-open-o pad-right-sm" />
+        {dir.name}
+      </span>
+      {indent === 0 && (
+        <div className="row">
+          <Button
+            variant="text"
+            title="Delete Directory"
+            onClick={event => {
+              event.stopPropagation();
+              handleDeleteDirectory(dir);
+            }}
+            bg="danger"
+          >
+            <i className="fa fa-trash-o" />
+          </Button>
+        </div>
+      )}
+    </ProtoListItem>),
   ...files.map(f => (
-    <ProtoFileListItem
+    <ProtoListItem
       key={f._id}
-      protoFile={f}
+      selectable
       isSelected={f._id === selectedId}
-      handleSelect={handleSelect}
-      handleUpdate={handleUpdate}
-      handleDelete={handleDelete}
-      indentLevel={indent}
-    />
+      onClick={() => handleSelect(f._id)}
+      indentLevel={indent + 1}
+    >
+      <>
+        <span className="wide">
+          <i className="fa fa-file-o pad-right-sm" />
+          {f.name}
+        </span>
+        <div className="row">
+          <Button
+            variant="text"
+            title="Re-upload Proto File"
+            onClick={event => {
+              event.stopPropagation();
+              handleUpdate(f);
+            }}
+            className="space-right"
+          >
+            <i className="fa fa-upload" />
+          </Button>
+          <Button
+            variant="text"
+            title="Delete Proto File"
+            bg="danger"
+            onClick={event => {
+              event.stopPropagation();
+              handleDelete(f);
+            }}
+          >
+            <i className="fa fa-trash-o" />
+          </Button>
+        </div>
+      </>
+    </ProtoListItem>
   )),
   ...subDirs.map(sd => recursiveRender(
-    indent,
+    indent + 1,
     sd,
     handleSelect,
     handleUpdate,
